// interfaces/course.ts
export type ICourse = {
  id: number
  title: string
  slug: string
  money_sign: string
  price: string
  discount: string
  teaser: string
  description: string
  more_info: string
  language: string
  level: string
  credit: any
  duration: string
  image_url: string
  video_url: string
  document_url: string
  is_featured: boolean
  created_at: string
  updated_at: string
  percent: string
  rating: string
  ratingNumber: number
  ratingCount: number
  email: string
  number: string
  bar: string
  category: string
  lessons_count: number
  instructorImg: string
  instructor: string
  categories: []
  authors: []
  enrol_users: []
  courseProgress: []
  lessons: []
  wishlist: []
  status: string
}